added custom decorator and static client page

+++ apps/nest-api/src/datasets/categories/categories.service.ts
@@ -1,10 +1,15 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { UpdateCategoryDto } from './dto/update-category.dto';
 import { PaginationDto } from 'src/common/custom-dto/pagination.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Category } from './entities/category.entity';
 import { Repository } from 'typeorm';
+import { DefaultPageSize } from 'src/utils/constants';
 
 @Injectable()
 export class CategoriesService {
@@ -17,7 +22,7 @@ export class CategoriesService {
     const { limit, offset } = paginationDto;
     return this.categoryRepository.find({
       take: limit,
-      skip: offset,
+      skip: offset ?? DefaultPageSize.CATEGORY,
     });
   }
 
@@ -29,6 +34,9 @@ export class CategoriesService {
     const category = await this.categoryRepository.findOne({
       where: {
         id,
+      },
+      relations: {
+        products: true,
       },
     });
 
@@ -50,6 +58,8 @@ export class CategoriesService {
 
   async remove(id: number) {
     const category = await this.findOne(id);
+    if (category.products.length)
+      throw new ConflictException('Category has related products');
     return this.categoryRepository.remove(category);
   }
 }
